refactor: removing nested onboarding directory/namespace
