+++ Makefile.env
@@ -1,8 +1,8 @@
 APP_NAME         = technical-on-boarding
 APP_VERSION      = 1.1.2
 APP_BUILD        = $(shell git rev-parse --short HEAD)
-APP_PACKAGE      = github.com/samsung-cnct/container-technical-on-boarding/onboarding
-APP_PATH         = ./onboarding/app
+APP_PACKAGE      = github.com/samsung-cnct/container-technical-on-boarding
+APP_PATH         = ./app
 APP_PATH_PKGS    = $(APP_PATH)/models $(APP_PATH)/controllers $(APP_PATH)/jobs $(APP_PATH)/jobs/onboarding
 
 GITHUB_ORG       = samsung-cnct
