--- onboarding/conf/routes
@@ -1,23 +1,0 @@
-# Routes Config
-#
-# This file defines all application routes (Higher priority routes first)
-#
-
-module:testrunner
-#module:jobs
-
-GET     /                                       App.Index
-GET     /version                                App.Version
-GET     /auth                                   App.Auth
-GET     /authcb                                 App.AuthCallback
-GET     /workload                               App.Workload
-WS      /workload/socket                        App.WorkloadSocket
-
-# Ignore favicon requests
-GET     /favicon.ico                            404
-
-# Map static resources from the /app/public folder to the /public path
-GET     /public/*filepath                       Static.Serve("public")
-
-# Catch all
-*       /:controller/:action                    :controller.:action
