--- onboarding/messages/sample.en
@@ -1,6 +1,0 @@
-# Sample messages file for the English language (en)
-# Message file extensions should be ISO 639-1 codes (http://en.wikipedia.org/wiki/List_of_ISO_639-1_codes)
-# Sections within each message file can optionally override the defaults using ISO 3166-1 alpha-2 codes (http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2)
-# See also:
-# - http://www.rfc-editor.org/rfc/bcp/bcp47.txt
-# - http://www.w3.org/International/questions/qa-accept-lang-locales
