--- onboarding/tests/apptest.go
@@ -1,29 +1,0 @@
-package tests
-
-import (
-	"github.com/revel/revel/testing"
-)
-
-type AppTest struct {
-	testing.TestSuite
-}
-
-func (t *AppTest) Before() {
-	println("Set up")
-}
-
-func (t *AppTest) TestThatIndexPageWorks() {
-	t.Get("/")
-	t.AssertOk()
-	t.AssertContentType("text/html; charset=utf-8")
-}
-
-func (t *AppTest) TestAppVersion() {
-	t.Get("/version")
-	t.AssertOk()
-	t.AssertContentType("application/json; charset=utf-8")
-}
-
-func (t *AppTest) After() {
-	println("Tear down")
-}
